drivers: npmx: shell: move buck code to the separate file
