+++ drivers/npmx/CMakeLists.txt
@@ -48,6 +48,7 @@ if(CONFIG_NPMX_SHELL)
     zephyr_library_sources(shell/shell_common.c)
 
     zephyr_library_sources(shell/adc.c)
+    zephyr_library_sources(shell/buck.c)
 endif()
 
 # Get npmx include file
